Add node and script for virtual marker
This commit add a node named add_markers, with
the whole purpose is to visualize the object to
be picked up and dropped off by the robot. It will
be later integrate with pick_objects node for fully
serviced home robot. All the coordination are hard
coded for now.

Also add a script to initialize all of these launch.

wiki http://wiki.ros.org/rviz/Tutorials/Markers%3A%20Basic%20Shapes