Create full home service robot
This commit attempts to create the first working
version of the home service robot, which takes
an object from a pick up position then drop off
to the desired position. This is done by rewrite
the add_markers node to listen to /amcl_pose and
check if the robot has reached the detination or not

Because of this change, the original add_markers code
is preserved in add_makers.cpp.orig. Beside, the rviz
configuration with the marker is stored in
home_service.rviz file, and the final script is
home_service.sh.